+++ packages/component-scanner/src/index.ts
@@ -1,7 +1,8 @@
 import { Listr } from 'listr2';
 import { chalk, path } from 'zx';
 import { type Visitor } from './ast-helper';
-import { callWorker, NamingStyle } from './worker';
+import { callWorker, type ComponentDetail } from './worker';
+import { type NamingStyle, transformNamingStyle } from './utils';
 
 /**
  * Export some utilities for conveniently using.
@@ -40,7 +41,7 @@ export interface ScanOptions {
    */
   libraryNames?: string[];
   /**
-   * This helps formatting the extracted component names.
+   * This helps to format the extracted component names.
    * @default 'default'
    */
   namingStyle?: NamingStyle;
@@ -48,12 +49,30 @@ export interface ScanOptions {
    * Whether to print log.
    * @default true
    */
-  verbose?: boolean
+  verbose?: boolean;
+  /**
+   * Sometimes you probably wrap a component from the library and use it in many places.
+   * You can set alias for these component so that their usage can be counted more exact.
+   * @example
+   * ```js
+   * {
+   *   alias: {
+   *     'custom-button': 'a-button', // The custom-button will be treated as a-button
+   *   }
+   * }
+   * ```
+   * The key and value will be transformed according to the namingStyle.
+   */
+  alias?: Record<string, string>;
+}
+
+interface Result extends ComponentDetail {
+  usage: Record<string, number>;
 }
 
 interface TaskCtx {
   files: string[];
-  results: string[];
+  results: Result[];
 }
 
 const globPattern = '**/*.{js,ts,jsx,tsx,vue,mjs,cjs}';
@@ -64,7 +83,7 @@ const logger = {
   error: (...args) => console.log(chalk.red('error'), ...args),
 };
 
-export async function scan(options?: ScanOptions) {
+export async function scanComponents(options?: ScanOptions) {
   let {
     cwd = process.cwd(),
     ignore = [],
@@ -73,16 +92,22 @@ export async function scan(options?: ScanOptions) {
     visitors,
     namingStyle,
     verbose = true,
+    alias = {},
   } = options || {};
 
   ignore = [...new Set([...ignore, '**/node_modules/**'])];
+  alias = Object.fromEntries(
+    Object.entries(alias).map(arr =>
+      arr.map(name => transformNamingStyle(name, namingStyle)),
+    ),
+  );
 
   if (verbose) {
     logger.info(`Working directory is ${chalk.underline.cyan(cwd)}`);
     logger.info(`Ignored paths: ${ignore.map(v => chalk.yellow(v)).join(',')}`);
   }
 
-  // Absolutize the paths.
+  // Make paths absolute.
   files = files.map(filename =>
     path.isAbsolute(filename)
       ? filename
@@ -126,7 +151,35 @@ export async function scan(options?: ScanOptions) {
           onProgress(filename, index, total) {
             task.output = `${index + 1}/${total} - ${filename}`;
           },
-        });
+        }) as Result[];
+      },
+    },
+    {
+      title: 'Applying alias...',
+      options: { bottomBar: 1 },
+      task: async (ctx, task) => {
+        const { length } = ctx.results;
+        const results: Result[] = [];
+        for (let i = 0; i < length; i++) {
+          const detail = ctx.results[i];
+          task.output = `${i + 1}/${length} - ${detail.filename}`; // print process
+          const usage: Result['usage'] = {};
+          for (let j = detail.components.length - 1; j >= 0; j--) {
+            const name = detail.components[j];
+            const realName = alias[name] ?? name; // replace aliased name with real name.
+            usage[realName] = (usage[realName] ?? 0) + 1;
+          }
+          // filter empty results
+          const names = Object.keys(usage);
+          if (names.length) {
+            results.push({
+              ...detail,
+              usage,
+              components: names,
+            });
+          }
+        }
+        ctx.results = results;
       },
     },
   ], {
@@ -142,3 +195,12 @@ export async function scan(options?: ScanOptions) {
 
   return ctx.results;
 }
+
+/**
+ * This function has been renamed to `scanComponents`, please use it instead.
+ * @deprecated
+ */
+export const scan: typeof scanComponents = (...args) => {
+  console.log(chalk.red('This function has been renamed to `scanComponents`, please use it instead.'));
+  return scanComponents.call(this, ...args);
+};