+++ packages/component-scanner/CHANGELOG.md
@@ -1,3 +1,8 @@
+# 1.1.0
+
+feat: add `alias` option to make analysis more exact.
+chore: exported function `scan` has been renamed to `scanComponents`.
+
 # 1.0.2
 
 docs: migrate to monorepo
@@ -8,4 +13,4 @@ refactor: adjust output structure, split worker as a independent file.
 
 # 1.0.0
 
-feat: Setup the project.
+feat: Set up the project.